practice_2/expected_output/pyproject.toml
[project]
name = "mcp-hotspots-handler"
version = "0.1.0"
description = "MCP tool handler for Git hotspot analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.46",
    "mcp>=1.26.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
